refactor(lab-pug-02):  strore
