+++ src/lab-pug-02/src/layout/layout.pug
@@ -2,7 +2,7 @@
 include ../store/commited-state.pug
 +load()
 - const exampleId='lab-pug-02'
-- locals.config = {}; locals.config.title= `Ciao, benvenuto in ${exampleId}`
+- locals.config = {}; locals.config.title= `::${exampleId}`
 
 // app layout
 doctype html
