+++ src/lab-pug-02/src/lib/actiontypes.js
@@ -1,0 +1,2 @@
+export const COUNT_UP = "COUNT_UP";
+export const COUNT_DOWN = "COUNT_DOWN";
