--- src/lab-pug-02/src/lib/state.js
@@ -1,7 +1,0 @@
-var state = {};
-
-export const getState = () => state;
-
-export const setState = nextState => {
-  state = nextState;
-};
